+++ config/initializers/form_builder.rb
@@ -1,0 +1,1 @@
+ActionView::Base.default_form_builder = ApplicationHelper::ValidationFormBuilder
